Race03
Matrix code, also known as Matrix digital rain or sometimes green rain, is the computercode related to the Matrix. The falling green code is a way of representing the activityof the virtual reality environment of the Matrix on screen."You get used to it, I don't even see the code. All I see is a blond, brunette, redhead."- Cypher